Update README.md to match the swagger docs
- Also refactored location of websocket handlers and created a dedicated file for HeaderSV handlers because the handlers.py file is getting too large.
