+++ esv_reference_server/handlers.py
@@ -42,77 +42,6 @@ async def error(request: web.Request) -> web.Response:
     raise ValueError("This is a test of raising an exception in the handler")
 
 
-async def get_header(request: web.Request) -> web.Response:
-    client_session: aiohttp.ClientSession = request.app['client_session']
-    app_state: ApplicationState = request.app['app_state']
-
-    accept_type = request.headers.get('Accept')
-    blockhash = request.match_info.get('hash')
-    if not blockhash:
-        return web.HTTPNotFound()
-
-    try:
-        url_to_fetch = f"{app_state.header_sv_url}/api/v1/chain/header/{blockhash}"
-        if accept_type == 'application/octet-stream':
-            request_headers = {'Content-Type': 'application/octet-stream'}  # Should be 'Accept'
-            async with client_session.get(url_to_fetch, headers=request_headers) as response:
-                result = await response.read()
-            response_headers = {'Content-Type': 'application/octet-stream', 'User-Agent': 'ESV-Ref-Server'}
-            return web.Response(body=result, status=200, reason='OK', headers=response_headers)
-
-        # else: application/json
-        request_headers = {'Content-Type': 'application/json'}  # Should be 'Accept'
-        async with client_session.get(url_to_fetch, headers=request_headers) as response:
-            result = await response.json()
-        response_headers = {'User-Agent': 'ESV-Ref-Server'}
-        return web.json_response(result, status=200, reason='OK', headers=response_headers)
-    except aiohttp.ClientConnectorError as e:
-        logger.error(f"HeaderSV service is unavailable on {app_state.header_sv_url}")
-        return web.HTTPServiceUnavailable()
-
-
-async def get_headers_by_height(request: web.Request) -> web.Response:
-    client_session: aiohttp.ClientSession = request.app['client_session']
-    app_state: ApplicationState = request.app['app_state']
-
-    accept_type = request.headers.get('Accept')
-    params = request.rel_url.query
-    height = params.get('height', '0')
-    count = params.get('count', '1')
-
-    try:
-        url_to_fetch = f"{app_state.header_sv_url}/api/v1/chain/header/byHeight?height={height}&count={count}"
-        if accept_type == 'application/octet-stream':
-            request_headers = {'Accept': 'application/octet-stream'}
-            async with client_session.get(url_to_fetch, headers=request_headers) as response:
-                result = await response.read()
-            response_headers = {'Content-Type': 'application/octet-stream', 'User-Agent': 'ESV-Ref-Server'}
-            return web.Response(body=result, status=200, reason='OK', headers=response_headers)
-
-        # else: application/json
-        request_headers = {'Accept': 'application/json'}
-        async with client_session.get(url_to_fetch, headers=request_headers) as response:
-            result = await response.json()
-        response_headers = {'User-Agent': 'ESV-Ref-Server'}
-        return web.json_response(result, status=200, reason='OK', headers=response_headers)
-    except aiohttp.ClientConnectorError as e:
-        logger.error(f"HeaderSV service is unavailable on {app_state.header_sv_url}")
-        return web.HTTPServiceUnavailable()
-
-
-async def get_chain_tips(request: web.Request) -> web.Response:
-    client_session: aiohttp.ClientSession = request.app['client_session']
-    app_state: ApplicationState = request.app['app_state']
-
-    url_to_fetch = f"{app_state.header_sv_url}/api/v1/chain/tips"
-    request_headers = {'Accept': 'application/json'}
-    async with client_session.get(url_to_fetch, headers=request_headers) as response:
-        result = await response.json()
-    response_headers = {'User-Agent': 'ESV-Ref-Server'}
-    # Todo - if we do not like json, we need to come up with a binary protocol for this
-    return web.json_response(result, status=200, reason='OK', headers=response_headers)
-
-
 async def get_account(request: web.Request) -> web.Response:
     """Two alternative forms of authentication. Either Bearer Token auth required"""
     app_state: ApplicationState = request.app['app_state']
